@@ -1,6 +1,6 @@
 {
   "name": "MYukkuriVoice",
-  "version": "0.9.6",
+  "version": "0.10.0",
   "description": "動画作成用ゆっくりボイス App for Mac OSX.",
   "main": "electron.js",
   "types": "types.d.ts",
@@ -32,7 +32,7 @@
   "license": "MIT",
   "dependencies": {
     "@types/angular": "^1.6.51",
-    "@types/ffi": "^0.2.0",
+    "@types/ffi": "^0.2.1",
     "@types/intro.js": "^2.4.3",
     "@types/mocha": "^5.2.5",
     "@types/node": "^10.11.7",
@@ -44,18 +44,23 @@
     "angular": "^1.7.5",
     "angular-input-highlight": "^0.1.2",
     "angular-ui-grid": "^4.6.4",
+    "bindings": "^1.3.1",
     "cryptico.js": "^1.2.0",
     "csv-parse": "^3.1.3",
     "csv-stringify": "^4.3.1",
+    "debug": "^4.1.1",
     "electron-json-storage": "^2.1.1",
     "electron-localshortcut": "^1.1.1",
     "electron-log": "^1.3.0",
     "electron-path": "git+https://github.com/taku-o/electron-path.git#master",
     "electron-store": "^2.0.0",
-    "ffi": "^2.2.0",
+    "ffi": "^2.3.0",
     "github-version-compare": "^0.1.0",
     "intro.js": "^2.9.3",
     "lru-cache": "^5.1.1",
+    "ms": "^2.1.1",
+    "nan": "^2.12.1",
+    "natives": "^1.1.6",
     "photon": "git+https://github.com/connors/photon.git",
     "ref": "^1.3.5",
     "ref-struct": "^1.1.0",
@@ -67,8 +72,9 @@
     "asar": "^0.14.5",
     "caller-position": "git+https://github.com/taku-o/caller-position.git#master",
     "devtron": "^1.4.0",
-    "electron": "^1.8.8",
-    "electron-packager": "^9.1.0",
+    "electron": "^3.1.1",
+    "electron-packager": "^12.1.0",
+    "electron-rebuild": "^1.8.2",
     "eslint": "^5.6.1",
     "fs-extra": "^7.0.1",
     "gulp": "^3.9.1",
@@ -90,7 +96,7 @@
     "node-notifier": "^5.2.1",
     "run-sequence": "^2.2.1",
     "source-map-support": "^0.5.9",
-    "spectron": "^3.8.0",
+    "spectron": "^5.0.0",
     "typescript": "^3.1.3",
     "typescript-eslint-parser": "^18.0.0",
     "yargs": "^12.0.2"
