@@ -33,7 +33,7 @@ const WORK_REPO_DIR = __dirname+ '/release/myukkurivoice';
 const WORK_UNPACK_DIR = __dirname+ '/release/myukkurivoice/MYukkuriVoice-darwin-x64/MYukkuriVoice.app/Contents/Resources/app.asar.unpacked';
 const APP_PACKAGE_NAME = 'MYukkuriVoice-darwin-x64';
 
-const ELECTRON_VERSION = '1.8.8';
+const ELECTRON_VERSION = '3.1.1';
 const APP_VERSION = require('./package.json').version;
 
 // default task
@@ -443,6 +443,7 @@ gulp.task('_package-release', (cb) => {
           --icon=icns/myukkurivoice.icns --overwrite --asar \
           --protocol-name=myukkurivoice --protocol=myukkurivoice \
           --extend-info=extend.plist \
+          --no-prune \
           --ignore="^/js/apps.spec.js" \
           --ignore="^/contents-spec.html" \
           --ignore="^/vendor" \
@@ -550,6 +551,7 @@ gulp.task('_package-release', (cb) => {
           --ignore="/LICENSE$" \
           --ignore="/LICENSE-MIT\\.txt$" \
           --ignore="/LICENSE-jsbn$" \
+          --ignore="/LICENSES\\.chromium\\.html$" \
           --ignore="/LICENSE\\.APACHE2$" \
           --ignore="/LICENSE\\.BSD$" \
           --ignore="/LICENSE\\.MIT$" \
@@ -589,11 +591,13 @@ gulp.task('_package-release', (cb) => {
           --ignore="/appveyor\\.yml$" \
           --ignore="/bower\\.json$" \
           --ignore="/component\\.json$" \
+          --ignore="/example\\.html$" \
           --ignore="/example\\.js$" \
           --ignore="/favicon\\.ico$" \
           --ignore="/gulpfile\\.js$" \
           --ignore="/karma\\.conf\\.js$" \
           --ignore="/license$" \
+          --ignore="/license\\.txt$" \
           --ignore="/package-lock\\.json$" \
           --ignore="/project\\.pbxproj$" \
           --ignore="/test\\.js$" \
@@ -720,6 +724,7 @@ gulp.task('_package-debug', (cb) => {
           --ignore="/LICENSE$" \
           --ignore="/LICENSE-MIT\\.txt$" \
           --ignore="/LICENSE-jsbn$" \
+          --ignore="/LICENSES\\.chromium\\.html$" \
           --ignore="/LICENSE\\.APACHE2$" \
           --ignore="/LICENSE\\.BSD$" \
           --ignore="/LICENSE\\.MIT$" \
@@ -759,11 +764,13 @@ gulp.task('_package-debug', (cb) => {
           --ignore="/appveyor\\.yml$" \
           --ignore="/bower\\.json$" \
           --ignore="/component\\.json$" \
+          --ignore="/example\\.html$" \
           --ignore="/example\\.js$" \
           --ignore="/favicon\\.ico$" \
           --ignore="/gulpfile\\.js$" \
           --ignore="/karma\\.conf\\.js$" \
           --ignore="/license$" \
+          --ignore="/license\\.txt$" \
           --ignore="/package-lock\\.json$" \
           --ignore="/project\\.pbxproj$" \
           --ignore="/test\\.js$" \
